// src/api-client/api.ts
export interface TaskListResponseTasksInner {
    /**
     * 
     * @type {TaskStatus}
     * @memberof TaskListResponseTasksInner
     */
    'status': TaskStatus;
    /**
     * The ID of the task
     * @type {number}
     * @memberof TaskListResponseTasksInner
     */
    'id': number;
    /**
     * The name of the task
     * @type {string}
     * @memberof TaskListResponseTasksInner
     */
    'name': string;
}


/**
 * The status of the task
 * @export
 * @enum {string}
 */

export const TaskStatus = {
    Running: 'running',
    Paused: 'paused'
} as const;

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];